Fixed partial coherence
The PartiallyCoherentScaledGaussianDataStack class seemed to be
incorrectly defined. First of all, it allowed values of any 'degree'
whereas the manual defines them as between 0 and 1. Secondly, it defined
the datastack as:

degree * incoherence + coherence

This leads to 0 being fully coherent (defined as fully incoherent in the
manual) and always lets the full coherent image as part of the field.
Changed this to:

(1 - degree) * incoherence + (degree * coherence)

This seems to be the computation that the manual describes, and also
what is implemented in the Matlab version. From "Setup_disarray.m":

% Add coherent and incoherent in parts defined by COHERENCE
RANDF_scale{nscale}=CohRANDF_scale*COHERENCE+inCohRANDF_scale*(1-COHERENCE);
% Multiply by reach value at the appropriate scale
RANDF_scale{nscale}=RANDF_scale{nscale}*REACH_SCALE(nscale);